Updated as per PR comments.
