+++ js-miniapp-sample/src/pages/landing.js
@@ -10,7 +10,6 @@ import LanguageRoundedIcon from '@mui/icons-material/LanguageRounded';
 import SettingsApplicationsRoundedIcon from '@mui/icons-material/SettingsApplicationsRounded';
 import StorageRoundedIcon from '@mui/icons-material/StorageRounded';
 import SystemUpdateIcon from '@mui/icons-material/SystemUpdate';
-import MobileHostAppIcon from '@mui/icons-material/SendToMobile';
 import Avatar from '@mui/material/Avatar';
 import List from '@mui/material/List';
 import ListItem from '@mui/material/ListItem';
@@ -103,16 +102,6 @@ const Landing = (props: LandingProps) => {
     }
   }, [props]);
 
-  useEffect(() => {
-    try {
-      if (document.readyState === 'complete') {
-        miniAppDidFinishLoad();
-      }
-    } catch (e) {
-      console.log(e);
-    }
-  }, []);
-
   function getDarkMode() {
     MiniApp.miniappUtils
       .isDarkMode()
@@ -121,17 +110,6 @@ const Landing = (props: LandingProps) => {
       })
       .catch((miniAppError) => {
         console.log('getDarkMode - Error: ', miniAppError);
-      });
-  }
-
-  function miniAppDidFinishLoad() {
-    MiniApp.miniappUtils
-      .miniAppFinishedLoading()
-      .then((response) => {
-        setHostNotified(true);
-      })
-      .catch((miniAppError) => {
-        console.log('miniAppFinishedLoading - Error: ', miniAppError);
       });
   }
 
@@ -244,17 +222,6 @@ const Landing = (props: LandingProps) => {
           <ListItemText
             primary="Build Type:"
             secondary={String(props.hostBuildType) || '-'}
-          />
-        </ListItem>
-        <ListItem>
-          <ListItemAvatar>
-            <Avatar>
-              <MobileHostAppIcon />
-            </Avatar>
-          </ListItemAvatar>
-          <ListItemText
-            primary="HostApp Notified:"
-            secondary={String(hostNotified)}
           />
         </ListItem>
         <ListItem>
